Refactor helpers to avoid cycles
